added more controllers, routes, safety middlewares

+++ src/controllers/sheet.controller.js
@@ -44,7 +44,15 @@ exports.getSheet = async (req, res) => {
 }
 exports.deleteSheet = async (req, res) => {
     try{
-        const {sheetId} = req.body;
+        const {sheetId, userId} = req.body;
+        if(!sheetId || !userId){
+            return response_400(res, "Missing required fields");
+        }
+        const user = await User.findByIdAndUpdate(
+            userId,
+            { $pull : { sheets : sheetId} },
+            { new : true }
+        );
         const sheet = await Sheet.findByIdAndDelete(sheetId);
         if(!sheet){
             return response_404(res, "Sheet not found");
@@ -58,7 +66,10 @@ exports.deleteSheet = async (req, res) => {
 }
 exports.updateSheet = async (req, res) => {
     try{
-        const { sheetId, name} = req.body;
+        const { sheetId, name } = req.body;
+        if(!sheetId || !name){
+            return response_400(res, "Missing required fields");
+        }
         const sheet = await Sheet.findByIdAndUpdate(sheetId, { name }, {new: true});
         if(!sheet){
             return response_404(res, "Sheet not found");
@@ -69,4 +80,39 @@ exports.updateSheet = async (req, res) => {
         console.log(error);
         return response_500(res, "Internal server error!");
     }
+}
+
+exports.cloneSheet = async (req, res) => {
+    try{
+        const { sheetId, userId, name } = req.body;
+        if(!sheetId || !userId || !name){
+            return response_400(res, "Missing required fields");
+        }
+        const user = await User.findById(userId);
+        if(!user){
+            return response_404(res, "User not found");
+        }
+
+        const sheet = await Sheet.findById(sheetId);
+        if(!sheet){
+            return response_404(res, "Sheet not found");
+        }
+
+        const clonedSheet = new Sheet({
+            name : name,
+            createdBy: userId,
+            problems : sheet.problems
+
+        })
+        await clonedSheet.save();
+
+        user.sheets.push(clonedSheet._id);
+        await user.save();
+
+        return response_200(res, "Sheet copied successfully", clonedSheet);
+        
+    } catch(error){
+        console.log(error);
+        return response_500(res, "Internal server error!");
+    }   
 }
